Feat: move search-header.tsx under src/components part of SWM-129 related to #45
